+++ test/normie_test.dart
@@ -1,1 +1,3 @@
+import 'package:normie/normie.dart';
+
 void main() {}